requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
psutil>=5.8.0
aiohttp>=3.8.0
asyncpg>=0.25.0
requests>=2.26.0